# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health/live": {
      "get": {
        "operationId": "HealthController_live",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "HealthController_ready",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "HealthController_index",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthPaginationDto"
                }
              }
            }
          }
        },
        "tags": [
          "Health"
        ]
      },
      "post": {
        "operationId": "HealthController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HealthDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Health"
        ]
      }
    },
    "/users/auth": {
      "post": {
        "operationId": "UserController_authorize",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizeUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizeUserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/auth/refresh": {
      "post": {
        "operationId": "UserController_refreshAuthTokens",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/password/change": {
      "post": {
        "operationId": "UserController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    }
  },
  "info": {
    "title": "Base Service API",
    "description": "REST API + GRPC endpoints",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "http://localhost:3132"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "totalPage": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPage"
        ]
      },
      "HealthDto": {
        "type": "object",
        "properties": {
          "test": {
            "type": "string",
            "description": "test string"
          }
        },
        "required": [
          "test"
        ]
      },
      "HealthPaginationDto": {
        "type": "object",
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HealthDto"
            }
          }
        },
        "required": [
          "meta",
          "items"
        ]
      },
      "AuthorizeUserDto": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "description": "Username or email",
            "example": "johndoe"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "P@ssw0rd123"
          }
        },
        "required": [
          "login",
          "password"
        ]
      },
      "AuthorizedUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "object",
            "nullable": true
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "role"
        ]
      },
      "AuthTokensDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "tokenType"
        ]
      },
      "AuthorizeUserResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthorizedUserDto"
          },
          "tokens": {
            "$ref": "#/components/schemas/AuthTokensDto"
          },
          "authenticatedAt": {
            "type": "string",
            "description": "Authorization timestamp in ISO format",
            "example": "2026-02-11T10:00:00.000Z"
          }
        },
        "required": [
          "user",
          "tokens",
          "authenticatedAt"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "RefreshTokenResponseDto": {
        "type": "object",
        "properties": {
          "tokens": {
            "$ref": "#/components/schemas/AuthTokensDto"
          },
          "refreshedAt": {
            "type": "string",
            "description": "Refresh timestamp in ISO format",
            "example": "2026-02-11T10:15:00.000Z"
          }
        },
        "required": [
          "tokens",
          "refreshedAt"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "example": "current_password"
          },
          "newPassword": {
            "type": "string",
            "example": "new_strong_password"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "ChangePasswordResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Password changed successfully"
          }
        },
        "required": [
          "message"
        ]
      }
    }
  }
}