# src/common/dto/pagination.dto.ts
import { ApiProperty } from '@nestjs/swagger';

class PaginationMeta {
  @ApiProperty()
  total: number;

  @ApiProperty()
  page: number;

  @ApiProperty()
  limit: number;

  @ApiProperty()
  totalPage: number;
}

export class PaginationDto {
  @ApiProperty()
  meta: PaginationMeta;
}